apps/api/pyproject.toml
[project]
name = "oculus-api"
version = "0.0.2"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "python-dotenv>=1.0.1",
  "supabase>=2.6.0",      # official supabase client (HTTP)
  "httpx>=0.27.2"
]
